feat: modernize mvp_dagger with Dagger 2.51, Room 2.6, ViewBinding and Firebase Firestore backend

+++ funnyvote/app/src/main/java/com/heaton/funnyvote/ui/WelcomeActivity.java
@@ -5,26 +5,27 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.heaton.funnyvote.FirstTimePref;
 import com.heaton.funnyvote.FunnyVoteApplication;
+import com.heaton.funnyvote.ui.mainactivity.MainActivity;
 import com.heaton.funnyvote.R;
 import com.heaton.funnyvote.data.user.UserDataRepository;
 import com.heaton.funnyvote.data.user.UserDataSource;
+import com.heaton.funnyvote.data.local.AppDatabase;
+import com.heaton.funnyvote.data.local.dao.PromotionDao;
 import com.heaton.funnyvote.database.Promotion;
-import com.heaton.funnyvote.database.PromotionDao;
 import com.heaton.funnyvote.database.User;
 import com.heaton.funnyvote.ui.introduction.IntroductionActivity;
-import com.heaton.funnyvote.ui.mainactivity.MainActivity;
 
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Created by heaton on 2016/10/26.
+ * Created by heaton on 2017/2/27.
  */
 
 public class WelcomeActivity extends AppCompatActivity {
@@ -33,8 +34,7 @@ public class WelcomeActivity extends AppCompatActivity {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_welcome);
-        InitTask task = new InitTask(this);
-        task.execute();
+        new InitTask(this).execute();
     }
 
     @Override
@@ -62,7 +62,8 @@ public class WelcomeActivity extends AppCompatActivity {
                 contextWeakReference.get().startActivity(
                         new Intent(contextWeakReference.get(), MainActivity.class));
             }
-            UserDataRepository userDataRepository = ((FunnyVoteApplication) (contextWeakReference.get().getApplication())).userDataRepository;
+            FunnyVoteApplication app = (FunnyVoteApplication) contextWeakReference.get().getApplication();
+            UserDataRepository userDataRepository = app.userDataRepository;
             userDataRepository.getUser(new UserDataSource.GetUserCallback() {
                 @Override
                 public void onResponse(User user) {
@@ -85,20 +86,21 @@ public class WelcomeActivity extends AppCompatActivity {
                 e.printStackTrace();
             }
             if (firstTimePref.getBoolean(FirstTimePref.SP_FIRST_MOCK_DATA, true)) {
-                PromotionDao promotionDao = ((FunnyVoteApplication) (contextWeakReference.get().getApplicationContext()))
-                        .getDaoSession().getPromotionDao();
-                String imageURL[] = contextWeakReference.get().getResources().getStringArray(R.array.imageURL);
-                List<Promotion> promotions = new ArrayList<>();
-                for (int i = 0; i < 1; i++) {
-                    Promotion promotion = new Promotion();
-                    promotion.setImageURL(imageURL[i % imageURL.length]);
-                    promotion.setActionURL("https://play.google.com/store/apps/details?id=com.heaton.funnyvote");
-                    promotion.setTitle("title:" + i);
-                    promotions.add(promotion);
+                if (contextWeakReference.get() != null) {
+                    PromotionDao promotionDao = AppDatabase.getInstance(contextWeakReference.get().getApplicationContext()).promotionDao();
+                    String imageURL[] = contextWeakReference.get().getResources().getStringArray(R.array.imageURL);
+                    List<Promotion> promotions = new ArrayList<>();
+                    for (int i = 0; i < 1; i++) {
+                        Promotion promotion = new Promotion();
+                        promotion.setImageURL(imageURL[i % imageURL.length]);
+                        promotion.setActionURL("https://play.google.com/store/apps/details?id=com.heaton.funnyvote");
+                        promotion.setTitle("title:" + i);
+                        promotions.add(promotion);
+                    }
+                    promotionDao.deleteAll();
+                    promotionDao.insertAll(promotions);
+                    firstTimePref.edit().putBoolean(FirstTimePref.SP_FIRST_MOCK_DATA, false).apply();
                 }
-                promotionDao.deleteAll();
-                promotionDao.insertInTx(promotions);
-                firstTimePref.edit().putBoolean(FirstTimePref.SP_FIRST_MOCK_DATA, false).apply();
             }
             return null;
         }
